Renamed and edited singleton class
