+++ app/src/main/java/com/mtlepberghenov/internship_playground/App.java
@@ -2,9 +2,9 @@ package com.mtlepberghenov.internship_playground;
 
 import android.app.Application;
 
-import com.mtlepberghenov.internship_playground.mvp.model.CarList;
+import com.mtlepberghenov.internship_playground.mvp.model.data.SingletonCarList;
 
 public class App extends Application {
 
-    CarList carList = CarList.init();
+    SingletonCarList singletonCarList = SingletonCarList.init();
 }
