+++ app/src/main/java/com/mtlepberghenov/internship_playground/mvp/model/CarMainModelImpl.java
@@ -1,5 +1,6 @@
 package com.mtlepberghenov.internship_playground.mvp.model;
 
+import com.mtlepberghenov.internship_playground.mvp.model.data.SingletonCarList;
 import com.mtlepberghenov.internship_playground.mvp.model.entity.Car;
 
 import java.util.ArrayList;
@@ -9,7 +10,7 @@ public class CarMainModelImpl implements MainModel {
     private ArrayList<Car> carList;
 
     public CarMainModelImpl() {
-        this.carList = CarList.init().getList();
+        this.carList = SingletonCarList.init().getList();
     }
 
     @Override
